converted convertToBuffer to sync
