+++ lib/helpers/index.js
@@ -4,7 +4,7 @@ module.exports = {
     buildPublishOrSendOptions   : require('./buildPublishOrSendOptions'),
     defaultConfiguration        : require('./defaultConfiguration'),
     createConnectionString      : require('./createConnectionString'),
-    convertToBufferAsync        : require('./convertToBufferAsync'),
+    convertToBuffer             : require('./convertToBuffer'),
     createTransactionId         : require('./createTransactionId'),
     exponentialBackoff          : require('./exponentialBackoff'),
     getPackageData              : require('./getPackageData'),
