+++ lib/index.js
@@ -362,14 +362,11 @@ class BunnyBus extends EventEmitter{
         const routeKey = Helpers.reduceRouteKey(null, options, message);
         const source = (options && options.source);
 
-        const [convertedMessage, transactionId, channelContext] = await Promise.all([
-            Helpers.convertToBufferAsync(message),
-            (async () => {
-
-                return (options && options.transactionId)
-                    ? options.transactionId
-                    : Helpers.createTransactionId();
-            })(),
+        const convertedMessage = Helpers.convertToBuffer(message);
+        const transactionId = (options && options.transactionId)
+            ? options.transactionId
+            : Helpers.createTransactionId();
+        const [channelContext] = await Promise.all([
             $._autoBuildChannelContext(BunnyBus.QUEUE_CHANNEL_NAME(queue), queue),
             $.createQueue(queue)
         ]);
@@ -440,14 +437,11 @@ class BunnyBus extends EventEmitter{
             throw new Exceptions.NoRouteKeyError();
         }
 
-        const [convertedMessage, transactionId, channelContext] = await Promise.all([
-            Helpers.convertToBufferAsync(message),
-            (async () => {
-
-                return (options && options.transactionId)
-                    ? options.transactionId
-                    : Helpers.createTransactionId();
-            })(),
+        const convertedMessage = Helpers.convertToBuffer(message);
+        const transactionId = (options && options.transactionId)
+            ? options.transactionId
+            : Helpers.createTransactionId();
+        const [channelContext] = await Promise.all([
             $._autoBuildChannelContext(BunnyBus.PUBLISH_CHANNEL_NAME()),
             $.createExchange(globalExchange, 'topic', null)
         ]);
